Work-in-progress: add gradle support for this project
Remove binaries, manage binaries through bintray instead.
We don't like this approach because it involves a poller, which is (a) a
single point of failure, and (b) another thing that can fail. We're going
to redo this using the cassandra reporting plugins rather than continuing
development here, but I figured I'd check in what I had for posterity.